WAITP-1317 Apply frontend excluded filter to all multi entity routes
