+++ packages/tupaia-web-server/src/routes/EntitySearchRoute.ts
@@ -6,6 +6,7 @@
 import { Request } from 'express';
 import { Route } from '@tupaia/server-boilerplate';
 import { TupaiaWebEntitySearchRequest } from '@tupaia/types';
+import { generateFrontendExcludedFilter } from '../utils';
 
 const DEFAULT_FIELDS = ['code', 'name', 'qualified_name'];
 
@@ -21,7 +22,16 @@ export class EntitySearchRoute extends Route<EntitySearchRequest> {
     const { projectCode } = params;
     const { searchString, page = 0, pageSize = 5, fields = DEFAULT_FIELDS } = query;
 
+    const project = (
+      await ctx.services.central.fetchResources('projects', {
+        filter: { code: projectCode },
+        columns: ['config'],
+      })
+    )[0];
+    const { config } = project;
+
     return ctx.services.entity.entitySearch(projectCode, searchString, {
+      filter: generateFrontendExcludedFilter(config),
       ...query,
       page,
       pageSize,
