+++ packages/tupaia-web-server/src/utils/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+export { generateFrontendExcludedFilter } from './generateFrontendExcludedFilter';
